Use existing identify event for signup attribution
